MDEV-4991: Fix wrong testcase binlog_encryption/encrypted_master_switch_to_unencrypted_gtid
This testcase actually demonstrates a bug in binlog encryption.
When encryption is enabled, non-encryption slaves are unable to replicate
the previous unencrypted binlogs using GTID, while this works correctly
without GTID.

For some reason, the testcase is written to expect this buggy behavior
rather than flag the bug. So fix the testcase to have the right expectation
now that this bug becomes fixed (from the gtid index code).

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>