MDEV-30660 COUNT DISTINCT seems unnecessarily slow when run on a PK
If we have a statement
  SELECT AGGR_FN(DISTINCT c1, c2,..,cn) FROM t1,
where AGGR_FN is an aggregate function like COUNT(), AVG(), SUM(),
and there is a unique index on table t1 including all columns (c1, c2,..,cn),
the values retrieved from the table are guaranteed to be unique.
So we can optimize aggregation and avoid de-duplicating of values,
in other words - ignore the DISTINCT clause.
The restrictions are:
  - only one table involved in the join
  - all arguments of the aggregate function are fields
        (not functions/subqueries)

Such an optimization can be also applied to cases
  SELECT AGGR_FN(c1, c2,..,cn) GROUP BY cx,..cy
when there is a unique index including all columns (c1, c2,..cn, cx,..cy)