MDEV-32899 instrumentation
In debug builds, let us declare dict_sys.latch as index_lock instead of
srw_lock, so that we will benefit from the full tracking of lock ownership.

lock_table_for_trx(): Assert that the current thread is not holding
dict_sys.latch. If the dict_sys.unfreeze() call were moved to the end of
lock_table_children(), this assertion would fail in the test innodb.innodb
and many other tests that use FOREIGN KEY.

dict_sys_t::lock(), dict_sys_t::freeze(): Assert that the current thread
is not holding dict_sys.latch in any mode. The SUX lock is reentrant,
but the simpler RW lock that we use in non-debug builds is not.